refactor map fragment
