+++ DDUBUCK/app/src/main/res/layout/activity_home.xml
@@ -6,9 +6,17 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".home.HomeActivity">
-
-    <include layout="@layout/map_fragment"
-        android:id="@+id/home_map_fragment"/>
+    <FrameLayout
+        android:id="@+id/home_map_container"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"/>
+    <FrameLayout
+        android:id="@+id/home_weather_score"
+        android:layout_marginTop="17dp"
+        android:layout_marginHorizontal="16dp"
+        android:layout_marginBottom="34dp"
+        android:layout_width="match_parent"
+        android:layout_height="100dp"/>
     <include layout="@layout/bottom_sheet"
         android:id="@+id/home_bottom_sheet"
         android:background="@android:color/white" />
